backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
requests
pydantic
python-dotenv
openai 
